requirement.txt
# python=3.12.7
# python -m pip install -r requirement.txt
torch --extra-index-url https://download.pytorch.org/whl/cu124 # specify your cuda
git+https://github.com/huggingface/peft.git
pytest
datasets
git+https://github.com/huggingface/transformers
accelerate
huggingface_hub
evaluate
scikit-learn
bitsandbytes
wandb
smart_open[s3]
python-dotenv
regex

# pip install ninja  # 确保有ninja加速编译
# git clone https://github.com/Dao-AILab/flash-attention.git
# cd flash-attention
# pip install . --no-build-isolation
